+++ meta_package_manager/labels.py
@@ -131,7 +131,7 @@ MANAGER_LABEL_GROUPS: TLabelGroup = {
     "asdf-based": frozenset({"asdf", "mise"}),
     "bash-based": frozenset({"basalt", "bpkg"}),
     "conda-based": frozenset({"conda", "mamba", "micromamba", "pixi"}),
-    "rpm-based": frozenset({"dnf", "dnf5", "urpmi", "yum", "zypper"}),
+    "rpm-based": frozenset({"dnf", "dnf5", "microdnf", "urpmi", "yum", "zypper"}),
     "dpkg-based": frozenset({
         "apt",
         "apt-mint",